% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_summary.R
\name{outbreak_summary}
\alias{outbreak_summary}
\title{Summarize scenario analysis}
\usage{
outbreak_summary(dat)
}
\arguments{
\item{res}{a list containing the case_data data.tables generated by scenario_sim}
}
\value{
data.table of outbreak summary statistics
}
\description{
Summarize scenario analysis
}
\author{
Ta-Chou Ng
}
