% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_outbreak_model.R
\name{pf_outbreak_model}
\alias{pf_outbreak_model}
\title{Run a single instance of the branching process model}
\usage{
pf_outbreak_model(
  num.initial.cases = NULL,
  cap_gen = NULL,
  r0isolated = NULL,
  r0community = NULL,
  disp.iso = NULL,
  disp.com = NULL,
  prop.asym = NULL,
  relR.asym = NULL,
  prop.ascertain = NULL,
  detect_sen = NULL,
  prop.presym = NULL,
  si_omega = NULL,
  incu_shape = NULL,
  incu_scale = NULL,
  delay_shape = NULL,
  delay_scale = NULL,
  delay_off = 0,
  quarant.days = NULL,
  quarant.retro.days = NULL
)
}
\arguments{
\item{r0isolated}{reproduction number for isolated cases (must be >0)}

\item{disp.iso}{dispersion parameter for isolated cases (must be >0)}

\item{disp.com}{dispersion parameter for non-isolated cases (must be >0)}

\item{prop.ascertain}{proportion of infectious contacts ascertained by contact tracing (must be 0<=x<=1)}

\item{delay_shape}{numeric shape parameter of delay distribution}

\item{delay_scale}{numeric scale parameter of delay distribution}
}
\value{
data.table of cases by week, cumulative cases, and the effective reproduction number of the outreak
}
\description{
Run a single instance of the branching process model
}
\examples{

\dontrun{
}
}
\author{
Joel Hellewell
}
