% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_outbreak_step.R
\name{pf_outbreak_step}
\alias{pf_outbreak_step}
\title{Move forward one generation in the branching process}
\usage{
pf_outbreak_step(
  case_data = NULL,
  r0symp = NULL,
  disp.com = NULL,
  disp.iso = NULL,
  r0isolated = NULL,
  prop.asym = NULL,
  relR.asym = NULL,
  prop.ascertain = NULL,
  quarant.days = NULL,
  quarant.retro.days = NULL,
  incfn = NULL,
  delayfn = NULL,
  inf_fn = NULL
)
}
\arguments{
\item{case_data}{data.table of cases in outbreak so far; initially generated by outbreak_setup}

\item{r0symp}{numeric reproduction number for non-isolated cases (must be >0)}

\item{disp.com}{numeric dispersion parameter for non-isolated cases (must be >0)}

\item{disp.iso}{numeric dispersion parameter for isolated cases (must be >0)}

\item{r0isolated}{numeric reproduction number for isolated cases (must be >0)}

\item{prop.ascertain}{numeric proportion of infectious contacts ascertained by contact tracing (must be 0<=x<=1)}

\item{quarant.days}{quarantine days}

\item{quarant.retro.days}{batracking of symptom onset}

\item{incfn}{function samples from incubation period; generated by dist_setup}

\item{delayfn}{function samples from the onset-to-hospitalisation delay; generated by dist_setup}

\item{k}{numeric skew parameter for sampling the serial interval from the incubation period}
}
\value{

}
\description{
Move forward one generation in the branching process
}
\author{
Joel Hellewell
}
