% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilter_fit.R
\name{pfilter_fit}
\alias{pfilter_fit}
\title{Particle filtering}
\usage{
pfilter_fit(
  model = NULL,
  fitted.paras = NULL,
  fixed.paras = NULL,
  measure = NULL,
  max.iter = 200,
  n.sam = 1000,
  eps = 0.1
)
}
\arguments{
\item{res}{a list containing the case_data data.tables generated by scenario_sim}
}
\value{
data.table of outbreak summary statistics
}
\description{
Particle filtering
}
\author{
Ta-Chou Ng
}
