% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilter_lhs.R
\name{pfilter_lhs}
\alias{pfilter_lhs}
\title{Sampling for particle filtering}
\usage{
pfilter_lhs(fitted.paras = NULL, n = NULL)
}
\arguments{
\item{res}{a list containing the case_data data.tables generated by scenario_sim}
}
\value{
data.table of outbreak summary statistics
}
\description{
Sampling for particle filtering
}
\author{
Ta-Chou Ng
}
